Remove several redundant python files in graph_net/ (#552)
* init 'symbolic_dimension_reifier' field in graph_net.json

* remove unused files

* add sample pass sample_pass/subgraph_input_shapes_naive_rewriter.py

* move graph_net_bench from ai4c

* remove duplicated samples

* remove several redundant python files in graph_net/

* minor fix
