+++ graph_net/graph_net_root.py
@@ -1,0 +1,6 @@
+import os
+import graph_net
+
+
+def get_graphnet_root():
+    return os.path.dirname(os.path.dirname(graph_net.__file__))
