+++ graph_net/paddle/test_reference_device.py
@@ -7,7 +7,7 @@ import json
 import sys
 import traceback
 
-from graph_net import path_utils
+from graph_net_bench import path_utils
 from graph_net import test_compiler_util
 from graph_net.paddle import test_compiler
