--- graph_net/path_utils.py
@@ -1,29 +1,0 @@
-import os
-import graph_net
-
-
-def get_graphnet_root():
-    return os.path.dirname(os.path.dirname(graph_net.__file__))
-
-
-def is_single_model_dir(model_dir):
-    return os.path.isfile(f"{model_dir}/graph_net.json") and os.path.isfile(
-        f"{model_dir}/model.py"
-    )
-
-
-def get_recursively_model_path(root_dir):
-    for sub_dir in get_immediate_subdirectory_paths(root_dir):
-        if is_single_model_dir(sub_dir):
-            yield sub_dir
-        else:
-            yield from get_recursively_model_path(sub_dir)
-
-
-def get_immediate_subdirectory_paths(parent_dir):
-    return [
-        sub_dir
-        for name in os.listdir(parent_dir)
-        for sub_dir in [os.path.join(parent_dir, name)]
-        if os.path.isdir(sub_dir)
-    ]
