+++ graph_net/torch/test_target_device.py
@@ -5,7 +5,7 @@ import sys
 import traceback
 
 import torch
-from graph_net import path_utils
+from graph_net_bench import path_utils
 from graph_net import test_compiler_util
 from graph_net.torch import test_compiler, test_reference_device
